update leetcode accepted submissions

+++ ContainsDuplicateIII.java
@@ -2,7 +2,7 @@ public class Solution {
     public boolean containsNearbyAlmostDuplicate(int[] nums, int k, int t) {
         TreeSet<Long> bst = new TreeSet();
         for (int i = 0; i < nums.length; ++ i) {
-            if (i > k) bst.remove(nums[i - k - 1]); // maintain k-size window
+            if (i > k) bst.remove((long) nums[i - k - 1]); // maintain k-size window
             long num = (long) nums[i];
             Long floor = bst.floor(num);
             Long ceil = bst.ceiling(num);
